Add FAQ in tree
In retrospect, it was a bad idea to use GitHub's wiki, which is itself
kind of a bad idea. I understand that it's supposed to support being a
real wiki in the sense that it can be edited by anyone, but considering
that GitHub requires you to commit even stuff like funding links and
issue templates in your code there's definitely no excuse for having the
wiki outside the codebase.